Moved perception utilities to their own file and cleaned up library imports
